Split lint pytest tests into files too
